feat: sui-429
After fighting with the serial console I finally got the runtimes and tools installed in usable locations. Firstly, you need the packAsTool directive so that it includes the necessary files to allow the dotnet cli to download it properly. Secondly, the environment we install to needs version 9 of the sdk. Additionally, we should specify the locations of our installs to ensure everythig is in path.