Inline PerformReconcilitation and restructure ReconcileAsync
Reqired to make sense of a bug where logically incorrect NHS number, provided as part of the request, is being used to retrieve demographics. Correct behaviour is: if the request's demographics do not match then reconcilitation should halt with a separate 'no match' status.

Audit logger was also removed, since this feature is now redundant (service is not expected to be hosted publically).